feat: Add comprehensive integration testing framework
- Create iDiski.Tests.Integration project with 23 end-to-end tests
- Test authentication flow: login, JWT generation, role claims (6 tests)
- Test team CRUD operations with authorization (7 tests)
- Test user management with SuperAdmin enforcement (10 tests)
- IntegrationTestFixture: WebApplicationFactory + in-memory database
- TestDataSeeder: Seeds test users with proper role assignments
- Both positive and negative test scenarios
- INTEGRATION_TESTING.md: Complete guide with patterns and best practices
- HTML test report with coverage matrix
- run-integration-tests.sh: Test runner script for CI/CD

Early issue detection enabled: Tests validate JWT parsing, database persistence,
authorization hierarchy, audit trail, and HTTP status codes end-to-end.

+++ iDiski.Infrastructure/Authorization/DivisionOwnershipHandler.cs
@@ -36,7 +36,7 @@ public class DivisionOwnershipHandler : AuthorizationHandler<DivisionOwnershipRe
 
         // Super Admin has full access
         var isSuperAdmin = await _context.UserRoles
-            .AnyAsync(ur => ur.UserId == userId && ur.Role == (int)Role.SuperAdmin);
+            .AnyAsync(ur => ur.UserId == userId && ur.Role == Role.SuperAdmin);
 
         if (isSuperAdmin)
         {
