+++ iDiski.Infrastructure/Authorization/TeamHierarchyHandler.cs
@@ -42,7 +42,7 @@ public class TeamHierarchyHandler : AuthorizationHandler<TeamOwnershipRequiremen
 
         // Super Admin has full access
         var isSuperAdmin = await _context.UserRoles
-            .AnyAsync(ur => ur.UserId == userId && ur.Role == (int)Role.SuperAdmin);
+            .AnyAsync(ur => ur.UserId == userId && ur.Role == Role.SuperAdmin);
 
         if (isSuperAdmin)
         {
